feat: add hash trait
and benefits from buildHash() and getHash() in to places now
